Use Import replace directly build Nacos Auth plugin bean in console.
