+++ naming/src/test/java/com/alibaba/nacos/naming/controllers/v2/OperatorControllerV2Test.java
@@ -26,7 +26,6 @@ import com.alibaba.nacos.naming.misc.SwitchDomain;
 import com.alibaba.nacos.naming.misc.SwitchManager;
 import com.alibaba.nacos.naming.model.form.UpdateSwitchForm;
 import com.alibaba.nacos.naming.model.vo.MetricsInfoVo;
-import com.alibaba.nacos.sys.env.Constants;
 import com.alibaba.nacos.sys.env.EnvUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -70,7 +69,6 @@ class OperatorControllerV2Test {
     void setUp() {
         this.operatorControllerV2 = new OperatorControllerV2(operatorV2Impl);
         MockEnvironment environment = new MockEnvironment();
-        environment.setProperty(Constants.SUPPORT_UPGRADE_FROM_1X, "true");
         EnvUtil.setEnvironment(environment);
     }
